enable contextIsolation in the renderer
